Change bandwidth storage from Int to Long.
Int is easy to overflow at just over 2 billion. As we're storing
bandwidth in bytes, and `/proc/net/dev` shows absolute number of bytes
transmitted since interface boot, it's easy to surpass the 2 billion
byte storage capacity of an int (~2TB of traffic is needed). Long can
support on the order of exabytes of traffic which, for this project,
should be plenty.

+++ src/main/kotlin/me/fru1t/lwnetmon/database/BandwidthDatabase.kt
@@ -9,5 +9,5 @@ interface BandwidthDatabase {
    *
    * @return the row id that is used as referenced internal to this database.
    */
-  fun addNextTick(deltaRxBytes: Int, deltaTxBytes: Int): Int
+  fun addNextTick(deltaRxBytes: Long, deltaTxBytes: Long): Int
 }
